mb got carried away icl
So far you can play the game. You start on the bottom left corner and the game correctly allows you to move in any direction you want. When you hit a wall you will be stuck in it until you enter the correct reversal move (so if you moved forwards into a wall you have to press backwards to get out)

The game also ends when you reached the end.

Few things that are missing: the output in the console is not identical to the brief, there is no maze boundary so you can fall out of the maze, the game doesn't keep track of the number of steps/mistakes